+++ pymetacode/templates/code/module.j2.py
@@ -1,4 +1,10 @@
 """
 {{ module.name }} module of the {{ package.name }} package.
 """
+{% if options.logging %}
+import logging
+
+
+logger = logging.getLogger(__name__)
+{% endif %}